requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.30
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
pymssql==2.2.10
cx_Oracle==8.3.0
anthropic==0.18.1
python-dotenv==1.0.1
requests==2.31.0
pytest==7.4.3
pandas==2.2.1
numpy==1.26.4
gunicorn==21.2.0 